+++ src/ion/__init__.py
@@ -1,3 +1,3 @@
 """ION - Documentation Template Management System."""
 
-__version__ = "0.39.8"
+__version__ = "0.39.9"